Update Blockchain.js
Added a dictionary called transactions_map, which is part of our data structure. The dictionary stores transactions in key-value pairs, where the keys are unique identifiers for each transaction, and the values are the transaction details.

When we add a new block to the blockchain, we loop through its transactions and add them to the transactions_map dictionary if they're not already there. Similarly, when we resolve conflicts and replace our chain with a longer one, we rebuild the entire transactions_map dictionary by looping through all transactions in all blocks in the chain and updating the dictionary accordingly.

## NEED TO FINE TUNE LATER ##